+++ srcs/containers/django/src/pong/static/js/classes/Player.js
@@ -1,0 +1,19 @@
+import { AI } from './AI.js';
+
+class Player {
+	constructor(name) {
+		this.name = name;
+		this.ai = null;
+
+		this.stats = {
+			wins: 0,
+			losses: 0,
+		};
+	}
+
+	setAI() {
+		this.ai = new AI();
+	}
+}
+
+export { Player };
