+++ srcs/containers/django/src/pong/static/js/main.js
@@ -1,7 +1,6 @@
 import Stats from './three-lib/stats.module.js'
 import { Game } from './classes/Game.js';
-import { Tournament } from './classes/Tournament.js';
-import { registerMatches } from './tournament-score-blockchain.js';
+import { loadPageClosure } from './loadPage.js';
 
 function main() {
 
@@ -28,151 +27,31 @@ function main() {
 	stats.showPanel(0);
 	document.body.appendChild(stats.dom);
 	
-	//key listener
-	let keys = {};
-	document.addEventListener("keydown", (event) => { 
-		keys[event.key] = true; 
-	});
-	
-	document.addEventListener("keyup", (event) => {
-		keys[event.key] = false;
-	});
-	
 	window.addEventListener('resize', () => this_game.onWindowResize());
 
 	// main loop
-	requestAnimationFrame(animate.bind(null, stats, keys, this_game));
+	requestAnimationFrame(animate.bind(null, stats, this_game));
 }
 
-function animate(stats, keys, game) {
-	stats.begin(); // for the FPS stats
-	game.update(keys);
+function animate(stats, game) {
+	// begin frame per second state
+	stats.begin();
+
+	// If the game is not running, render the menu view, 
+	// Else, update the game
+	if (game.running === false) {
+		const split = document.getElementById('vertical-line');
+		split.style.display = 'none';
+		game.cam1.renderMenuView(game);
+	} else {
+		game.match.update();
+	}
+
 	game.controls.update();
-	requestAnimationFrame(animate.bind(null, stats, keys, game));
-	stats.end(); // for the FPS stats
-}
+	requestAnimationFrame(animate.bind(null, stats, game));
 
-function fadeIn(element) {
-	return new Promise((resolve) => {
-		element.style.opacity = 0;
-		element.style.display = 'block';
-		element.offsetHeight; 
-		element.style.transition = 'opacity 1s';
-		element.style.opacity = 1;
-		element.addEventListener('transitionend', function handler() {
-			element.removeEventListener('transitionend', handler);
-			resolve();
-		}, { once: true });
-	});
-}
-
-function fadeOut(element) {
-	return new Promise((resolve) => {
-		element.style.transition = 'opacity 1s';
-		element.style.opacity = 0;
-		element.addEventListener('transitionend', function handler() {
-			element.removeEventListener('transitionend', handler);
-			resolve();
-		}, { once: true });
-	});
-}
-
-function loadPageClosure(game) {
-	return async (page) => {
-		try {
-			const mainContent = document.getElementById('main-content');
-			const underTitle = document.getElementById('under-title');
-			
-			if (underTitle) {
-				await fadeOut(underTitle);
-			}
-			
-			const response = await fetch('/api/' + page + '/');
-			if (!response.ok) {
-				throw new Error(`HTTP error! Status: ${response.status}`);
-			}
-			
-			const data = await response.json();
-			mainContent.innerHTML = data.content;
-			history.pushState({ page: page }, "", "#" + page);
-			
-			const updatedUnderTitle = document.getElementById('under-title');
-			if (updatedUnderTitle) {
-				await fadeIn(updatedUnderTitle);
-			}
-			bindEventListeners(game);
-			
-		} catch (error) {
-			console.error('Error loading page:', error);
-			throw error;
-		}
-	};
-}
-
-function bindEventListeners(game) {
-
-	const leaderBoardButton = document.getElementById('js-leaderboard-btn');
-	if (leaderBoardButton) {
-		leaderBoardButton.addEventListener('click', showLeaderBoard);
-	}
-
-	const leaderBoardClose = document.getElementById('js-leaderboard-close');
-	if (leaderBoardClose) {
-		leaderBoardClose.addEventListener('click', hideLeaderBoard);
-	}
-
-	const startUserVsUserButton = document.getElementById('js-start-user-vs-user-btn');
-	if (startUserVsUserButton) {
-		startUserVsUserButton.addEventListener('click', game.initMatch.bind(game, 'user-vs-user'));
-	}
-
-	const addPlayerBtn = document.getElementById('js-add-player-btn');
-	if (addPlayerBtn) {
-		var tournament = new Tournament(game);
-		addPlayerBtn.addEventListener('click', tournament.addPlayer.bind(tournament));
-	}
-	
-	const gameSoloBtn = document.getElementById('js-start-game-solo-btn');
-	if (gameSoloBtn) {
-		gameSoloBtn.addEventListener('click', game.initMatch.bind(game, 'user-vs-computer'));
-	}
-	
-	let startTournamentBtn = document.getElementById('js-start-tournament-btn');
-	if (startTournamentBtn) {
-		startTournamentBtn.addEventListener('click', tournament.startTournament.bind(tournament));
-	}
-
-	startTournamentBtn = document.getElementById('js-end-game-btn');
-	if (startTournamentBtn) {
-	  startTournamentBtn.addEventListener('click', game.endGame.bind(game));
-	}
-
-	var blockchainScore = document.getElementById('js-register-blockchain');
-	if (blockchainScore) {
-		blockchainScore.addEventListener('click', registerMatches.bind(null, game));
-	}
-}
-
-function showLeaderBoard() {
-
-	const leaderBoardDiv = document.getElementById('js-leaderboard-div');
-	if (leaderBoardDiv)
-		leaderBoardDiv.style.display = 'block';
-
-	const leaderBoardButton = document.getElementById('js-leaderboard-btn');
-	if (leaderBoardButton)
-		leaderBoardButton.style.display = 'none';
-}
-
-function hideLeaderBoard() {
-
-	const leaderBoardDiv = document.getElementById('js-leaderboard-div');
-	if (leaderBoardDiv)
-		leaderBoardDiv.style.display = 'none';
-
-	const leaderBoardButton = document.getElementById('js-leaderboard-btn');
-	if (leaderBoardButton)
-		leaderBoardButton.style.display = 'block';
+	// End the frame per second stats
+	stats.end(); 
 }
 
 main();
