--- srcs/containers/django/src/pong/static/js/tournament-end.js
@@ -1,3 +1,0 @@
-
-
-export {endTournament};
